Use transformers Grounding Dino to replace third party implementation
